requirements.txt
Django==5.2
python-dotenv==1.1.0
django-environ==0.12.0
PyMySQL==1.1.1
stripe==12.0.0
Pillow==11.1.0
uuid==1.30
django-crispy-forms==2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-stripe-payments==2.0.0
django-widget-tweaks==1.5.0
requests==2.32.3
pydantic==2.11.2
gunicorn==23.0.0
python-decouple==3.8
whitenoise==6.9.0
